feat: Unify jQUery Templates. Additionally rename $(gridfeatures) to $(treeGridFeatures).
