+++ templates/jquery/js/grid/grid-editing/files/index.html
@@ -29,6 +29,7 @@
                width: "500px",
                dataSource: products,
 			   primaryKey: "ProductID",
+			   autoCommit : true,
                features: $(gridfeatures)
            });
        });
