+++ templates/jquery/js/hierarchical-grid/hierarchical-grid-custom/files/index.html
@@ -15,6 +15,7 @@
 		 <script>
 		   $(function () {
 			$("#grid").igHierarchicalGrid({
+				autoCommit : true,
 				width: "100%",
 				autoGenerateColumns: false,
 				dataSource: northwind,
