+++ templates/jquery/js/hierarchical-grid/hierarchical-grid-editing/files/index.html
@@ -13,6 +13,7 @@
 			$(function() {
 	
 				$("#hierarchicalGrid").igHierarchicalGrid({
+					autoCommit : true,
 					width: "100%",
 					autoGenerateColumns: false,
 					dataSource: northwind,
