+++ templates/jquery/js/tree-grid/file-explorer/files/Index.html
@@ -65,7 +65,9 @@
                 {
                     name: "Paging",
                     pageSize: 4
-                }]
+				},
+				$(treeGridFeatures)
+			]
             });
 
 	});
