+++ templates/jquery/js/tree-grid/tree-grid-custom/files/index.html
@@ -26,7 +26,7 @@
 			   ],
 			   childDataKey: "products",
 			   features: [
-				 $(Gridfeatures)
+				 $(treeGridFeatures)
 			   ]
 		   });
        });
