+++ templates/jquery/js/tree-grid/tree-grid-editing/files/index.html
@@ -45,7 +45,7 @@
 								editorType: "datepicker"
 							}
 						]
-				   }$(Gridfeatures)
+				   }$(treeGridFeatures)
 			   ]
 		   });
        });
