+++ templates/jquery/js/tree-grid/tree-grid-export/files/index.html
@@ -43,7 +43,7 @@
                     { headerText: "Reports To", key: "reportsTo", width: "130px", dataType: "number" }
 				],
 				features: [
-					$(gridfeatures)
+					$(treeGridFeatures)
 				]
             });
 
@@ -65,7 +65,7 @@
                 initialExpandDepth: 1,
 				renderExpansionIndicatorColumn: true,
 				features: [
-					$(gridfeatures)
+					$(treeGridFeatures)
 				]
             });
